requirements.txt
loguru
mkdocs
pip
python-dotenv
ruff
tqdm
typer
pandas
numpy<2
scikit-surprise
scikit-learn
matplotlib
seaborn
jupyter
joblib
-e .
